Enabled PIC, got rid of size_t

+++ src/main/lp/sleqp_lpi.c
@@ -40,8 +40,8 @@ SLEQP_RETCODE sleqp_lpi_create_interface(SleqpLPi** lp_star,
 }
 
 SLEQP_RETCODE sleqp_lpi_create_problem(SleqpLPi* lp_interface,
-                                       size_t num_variables,
-                                       size_t num_constraints)
+                                       int num_variables,
+                                       int num_constraints)
 {
   assert(lp_interface);
 
@@ -70,7 +70,7 @@ SLEQP_RETCODE sleqp_lpi_solve(SleqpLPi* lp_interface,
 }
 
 SLEQP_RETCODE sleqp_lpi_get_solution(SleqpLPi* lp_interface,
-                                     size_t num_variables,
+                                     int num_variables,
                                      double* objective_value,
                                      double* solution_values)
 {
@@ -83,7 +83,7 @@ SLEQP_RETCODE sleqp_lpi_get_solution(SleqpLPi* lp_interface,
 }
 
 SLEQP_RETCODE sleqp_lpi_get_varstats(SleqpLPi* lp_interface,
-                                     size_t num_variables,
+                                     int num_variables,
                                      SLEQP_BASESTAT* variable_stats)
 {
   assert(lp_interface);
